+++ locksmith/__tests__/controllers/userController/userCreation.test.ts
@@ -1,10 +1,10 @@
 import { ethers } from 'ethers'
 
-import models = require('../../../src/models')
-import app = require('../../../src/app')
-import Base64 = require('../../../src/utils/base64')
+const models = require('../../../src/models')
+const app = require('../../../src/app')
+const Base64 = require('../../../src/utils/base64')
 
-const UserOperations = require('../../../src/operations/userOperations')
+import UserOperations from '../../../src/operations/userOperations'
 
 function generateTypedData(message: any, messageKey: string) {
   return {