+++ unlock-app/src/components/interface/keychain/Key.tsx
@@ -212,10 +212,6 @@ const Key = ({ ownedKey, account, network }: Props) => {
     }
   }
 
-  const onCancelAndRefund = () => {
-    setShowCancelModal(true)
-  }
-
   const iconButtonClass =
     'flex items-center disabled:opacity-50 disabled:border-gray-200 disabled:cursor-not-allowed p-2 border border-gray-100 rounded shadow opacity-90 hover:opacity-100 hover:border-gray-200'
   const sendEmail = (recipient: string, qrImage: string) => {
@@ -256,12 +252,13 @@ const Key = ({ ownedKey, account, network }: Props) => {
         keyId={keyId}
         network={network}
       />
+
       {!isKeyExpired && (
         <CancelAndRefundModal
-          active={showCancelModal}
+          isOpen={showCancelModal}
+          setIsOpen={setShowCancelModal}
           lock={lock}
           keyId={keyId}
-          setIsOpen={setShowCancelModal}
           account={account}
           currency={symbol}
           network={network}
@@ -272,7 +269,8 @@ const Key = ({ ownedKey, account, network }: Props) => {
       {signature && (
         <QRModal
           lock={lock}
-          active={showingQR}
+          isOpen={showingQR}
+          setIsOpen={setShowingQR}
           dismiss={() => setSignature(null)}
           sendEmail={sendEmail}
           signature={signature}
@@ -341,7 +339,7 @@ const Key = ({ ownedKey, account, network }: Props) => {
                 className={iconButtonClass}
                 type="button"
                 disabled={wrongNetwork}
-                onClick={onCancelAndRefund}
+                onClick={() => setShowCancelModal(!showCancelModal)}
               >
                 <CancelIcon />
               </button>