+++ unlock-app/src/hooks/useLocks.js
@@ -14,6 +14,7 @@ import { ConfigContext } from '../utils/withConfig'
 import { processTransaction } from './useLock'
 import { useConfig } from '~/utils/withConfig'
 import GraphService from '~/services/graphService'
+import { SubgraphService } from '@unlock-protocol/unlock-js'
 
 /**
  * Retrieves a lock object at the address
@@ -22,7 +23,6 @@ export const getLockAtAddress = async (web3Service, address, network) => {
   let lock
   try {
     lock = await web3Service.getLock(address, network)
-    lock.unlimitedKeys = lock.maxNumberOfKeys === UNLIMITED_KEYS_COUNT
     lock.address = address
     lock.network = network
   } catch (error) {
@@ -46,11 +46,25 @@ export const retrieveLocks = async (
   // The locks from the subgraph miss some important things, such as balance,
   // ERC20 info.. etc so we need to retrieve them from unlock-js too.
   // TODO: add these missing fields to the graph!
-  const locks = await graphService.locksByManager(owner, network)
+
+  // ignore localhost
+  if (network == 31337) return []
+  const service = new SubgraphService()
+  const locks = await service.locks(
+    {
+      first: 1000,
+      where: {
+        lockManagers_contains: [owner],
+      },
+    },
+    {
+      networks: [`${network}`],
+    }
+  )
 
   // Sort locks to show the most recent first
   locks.sort((x, y) => {
-    return parseInt(y.creationBlock) - parseInt(x.creationBlock)
+    return parseInt(y.createdAtBlock) - parseInt(x.createdAtBlock)
   })
 
   const loadNext = async (locks, done) => {